200611 project(오류 수정, 검색, header, footer 구현)

+++ workspace_jsp/Team2Project/src/member/command/DeleteCommand.java
@@ -8,6 +8,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import member.dao.MemberDAO;
+import member.domain.MemberDTO;
 import share.Command;
 import share.CommandAction;
 
@@ -16,15 +17,27 @@ public class DeleteCommand implements Command {
 	@Override
 	public CommandAction execute(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-	    HttpSession session = request.getSession(false);
-		
-		String id = request.getParameter("id");
-		
-		MemberDAO dao = new MemberDAO();
-		dao.delete(id);
-		session.invalidate();
-	
-		return new CommandAction(true, "memberselect.do");
+
+		HttpSession session = request.getSession(false);
+		MemberDTO dto = null;
+		if (session != null) {
+			dto = (MemberDTO) session.getAttribute("login");			
+			if (dto != null) {
+				String id = request.getParameter("id");
+				if (dto.getId().equals(id)) {
+					MemberDAO dao = new MemberDAO();
+					dao.delete(id);
+					session.invalidate();
+					return new CommandAction(true, "reviewlist.do");
+				} else {
+					return new CommandAction(true, "memberloginui.do");
+				}			
+			} else {
+				return new CommandAction(true, "memberloginui.do");
+			}
+		} else {
+			return new CommandAction(true, "memberloginui.do");
+		}
 	}
 
 }
